requirements.txt
streamlit
plotly
pandas
numpy
scipy
librosa
soundfile
pydub
scikit-learn
openai
python-dotenv
reportlab
ffmpeg-python
